modified

+++ master-slave/master/round2d.ino
@@ -1,4 +1,4 @@
-float round2d(float val)
+double round2d(float val)
 {
   return (int)(val * 100 + 0.5) / 100.0;
 }
